data_tools/0_shared/pyproject.toml
[project]
name = "0-shared"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pandas>=3.0.2",
    "psycopg2-binary>=2.9.12",
    "python-dotenv>=1.2.2",
    "sqlalchemy>=2.0.49",
    "sqlmodel>=0.0.38",
    "unidecode>=1.4.0",
]
